requirements.txt
pytest
pylightning
bitcoin-requests
